chapter_001/src/main/java/ru/job4j/sort/Merge.java
package ru.job4j.sort;

import java.util.Arrays;

/**
 * Program merges elements of two sorted arrays to a third sorted array.
 *
 * @author Rustem Mukazhanov (r.mukazhanov@gmail.com)
 * @version $Id$
 * @since 0.1
 */
public class Merge {
    /**
     * Method finds smallest element and stores it in sorted manner.
     *
     * @param left  first array.
     * @param right second array.
     * @return third array.
     */
    public int[] merge(int[] left, int[] right) {
        int[] rsl = new int[left.length + right.length];
        int i = 0, j = 0, k = 0;
        while (i < left.length && j < right.length) {
            if (left[i] < right[j]) {
                rsl[k++] = left[i++];
            } else {
                rsl[k++] = right[j++];
            }
        }
        while (i < left.length) {
            rsl[k++] = left[i++];
        }
        while (j < right.length) {
            rsl[k++] = right[j++];
        }
        return rsl;
    }

    /**
     * Main method prints out sorted array.
     *
     * @param args command line arguments.
     */
    public static void main(String[] args) {
        Merge process = new Merge();
        int[] rsl = process.merge(
                new int[]{1, 3, 5},
                new int[]{2, 4}
        );
        System.out.println(Arrays.toString(rsl));
    }
}
